LIbrary processor
Processor, which using libraries for work (regexp, pigeon, govaluate). Now the best working (if don't take into that there is no support for the construction of a number to degree, anything else - no problems).
At the moment, he immediately determines “damaged” expressions and does not consider them expressions. Therefore, temporarily tests of this processor are separate from the rest (it is planned to add this support to other processors)